refactor(downloads): 拆分DownloadsPage为组件选择器+实例配置面板
